Module Config.Client_invocation

type t =

| Connect_to of {

connection_config : Grpc_discovery.Connection_config.t;

}

| Offline

In the tests, sometimes we run client commands that needs to connect to the server, and sometimes we run client commands that do not, and only perform other kinds of actions. This type is used to distinguish between these cases.