Module Config.Process_command

type t = {

executable : Base.string;

args : Base.string Base.list;

}

This type is used to represent a command that the library must run to start the servers and clients processes. executable must be the path to the executable to run, searching $PATH for it if necessary.

Do not forget to list the executable as a test dependency in your dune file. For example, if you executable is "m_app", you should have something like this in your dune file:

          (library
            (name my_app_test)
            (inline_tests (deps %{bin:my_app}))
          ...)